requirements.txt
python-dotenv
python-telegram-bot==13.7
requests
